+++ api/tests/integration/edge_api/identities/test_edge_identity_viewset.py
@@ -1,9 +1,7 @@
 import json
 import urllib
 from typing import Any
-from unittest.mock import Mock
 
-import pytest
 from boto3.dynamodb.conditions import Key
 from django.urls import reverse
 from mypy_boto3_dynamodb.service_resource import Table
@@ -116,52 +114,6 @@ def test_create_identity(  # type: ignore[no-untyped-def]
     assert response.status_code == status.HTTP_201_CREATED
     assert response.json()["identifier"] == identifier
     assert response.json()["identity_uuid"] is not None
-
-
-@pytest.mark.valid_identity_identifiers
-@pytest.mark.usefixtures(
-    "dynamo_enabled_environment",
-)
-def test_create_identity_accepts_valid_identifiers(
-    admin_client: APIClient,
-    environment_api_key: str,
-    edge_identity_dynamo_wrapper_mock: Mock,
-    identifier: str,
-) -> None:
-    # Given
-    edge_identity_dynamo_wrapper_mock.get_item.return_value = None
-
-    # When
-    response = admin_client.post(
-        f"/api/v1/environments/{environment_api_key}/edge-identities/",
-        data={"identifier": identifier},
-    )
-
-    # Then
-    assert response.status_code == status.HTTP_201_CREATED
-    assert response.json()["identifier"] == identifier
-
-
-@pytest.mark.invalid_identity_identifiers
-@pytest.mark.usefixtures(
-    "dynamo_enabled_environment",
-    "edge_identity_dynamo_wrapper_mock",
-)
-def test_create_identity_responds_400_if_identifier_is_invalid(
-    admin_client: APIClient,
-    environment_api_key: str,
-    identifier: str,
-    identifier_error_message: str,
-) -> None:
-    # When
-    response = admin_client.post(
-        f"/api/v1/environments/{environment_api_key}/edge-identities/",
-        data={"identifier": identifier},
-    )
-
-    # Then
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"identifier": [identifier_error_message]}
 
 
 def test_create_identity_returns_400_if_identity_already_exists(  # type: ignore[no-untyped-def]
